Add Favs Logic
Update api calls to add favorite, remove favorite, get favorites from the Database. Card.js Toggle favorites method retrieves current user favorites array and compares
the movie ids from that array to the selected favorite movie. If the id does not match then movie is sent to the DB via addFavorites apiCall. If
id does match, then removeFavorite is called and that movie is removed from the Favorites array on the DB. Users logout functioanlity also works.
If the user is logged in, then login/sign in will show Logout. When logout is clicked, the store is updated to an empty object that replaces the current user.