Add local storage to remember user
Update ApiCall.js to create a returning user method that takes in the ID from local storage
and compares it to the returned user array from the users DB. ReturningUser method is called on
component did mount in APP.js if there is a UserId in the local storage. The returned user
is then added to the store via getUser redux prop.